feat: adding support to invoke via nodejs apps
