+++ packages/cli/src/scanner/ai-scanner.ts
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { Spinner } from 'cli-spinner';
+
 import { inject, injectable } from 'inversify';
 import * as util from 'util';
 import { AxeScanResults } from './axe-scan-results';
@@ -11,11 +11,9 @@ export class AIScanner {
     constructor(@inject(Page) private readonly page: Page) {}
 
     public async scan(url: string): Promise<AxeScanResults> {
-        const spinner = new Spinner();
         try {
             console.log(`Starting accessibility scanning of URL ${url}.`);
 
-            spinner.start();
             await this.page.create();
             await this.page.enableBypassCSP();
 
@@ -25,7 +23,6 @@ export class AIScanner {
 
             return { error: util.inspect(error) };
         } finally {
-            spinner.stop();
             await this.page.close();
             console.log(`Accessibility scanning of URL ${url} completed.`);
         }
