+++ packages/cli/webpack.config.js
@@ -2,19 +2,14 @@
 // Licensed under the MIT License.
 const path = require('path');
 const webpack = require('webpack');
-
+const nodeExternals = require('webpack-node-externals');
 const ForkTsCheckerWebpackPlugin = require('fork-ts-checker-webpack-plugin');
 const copyWebpackPlugin = require('copy-webpack-plugin');
+const { CheckerPlugin } = require('awesome-typescript-loader');
 
-module.exports = env => {
-    const version = env ? env.version : 'dev';
-    console.log(`Building for version : ${version}`);
+function getCommonConfig(version, generateTypings) {
     return {
         devtool: 'cheap-source-map',
-        externals: ['puppeteer', 'yargs', 'axe-core', 'axe-puppeteer'],
-        entry: {
-            ['ai-scan']: path.resolve('./src/index.ts'),
-        },
         mode: 'development',
         module: {
             rules: [
@@ -22,9 +17,9 @@ module.exports = env => {
                     test: /\.ts$/,
                     use: [
                         {
-                            loader: 'ts-loader',
+                            loader: 'awesome-typescript-loader',
                             options: {
-                                transpileOnly: true,
+                                transpileOnly: generateTypings ? false : true, // transpileOnly=false generates typings
                                 experimentalWatchApi: true,
                             },
                         },
@@ -45,21 +40,13 @@ module.exports = env => {
                 },
             ],
         },
-        name: 'ai-scan',
-        node: {
-            __dirname: false,
-        },
-        output: {
-            path: path.resolve('./dist'),
-            filename: '[name].js',
-            libraryTarget: 'commonjs2',
-        },
         plugins: [
+            new CheckerPlugin(),
             new webpack.BannerPlugin({ banner: '#!/usr/bin/env node', raw: true }),
             new webpack.DefinePlugin({
                 __IMAGE_VERSION__: JSON.stringify(version),
             }),
-            new ForkTsCheckerWebpackPlugin(),
+
             new copyWebpackPlugin([
                 {
                     context: './run-script',
@@ -68,11 +55,46 @@ module.exports = env => {
                     ignore: ['dist/**', 'node_modules/**'],
                 },
             ]),
-        ],
+        ].concat(
+            generateTypings
+                ? [] // add type checking if transpileOnly is set to false
+                : new ForkTsCheckerWebpackPlugin(),
+        ),
         resolve: {
             extensions: ['.ts', '.js', '.json'],
             mainFields: ['main'], //This is fix for this issue https://www.gitmemory.com/issue/bitinn/node-fetch/450/494475397
         },
         target: 'node',
+        node: {
+            __dirname: false,
+        },
+        output: {
+            path: path.resolve('./dist'),
+            filename: '[name].js',
+            libraryTarget: 'umd',
+        },
     };
+}
+
+module.exports = env => {
+    const version = env ? env.version : 'dev';
+    console.log(`Building for version : ${version}`);
+    return [
+        {
+            ...getCommonConfig(version, false),
+            name: 'ai-scan-cli',
+            externals: ['puppeteer', 'yargs', 'axe-core', 'axe-puppeteer'],
+            entry: {
+                ['ai-scan-cli']: path.resolve('./src/cli.ts'),
+            },
+        },
+        {
+            ...getCommonConfig(version, true),
+            name: 'ai-scan',
+            externals: [nodeExternals()],
+            entry: {
+                ['index']: path.resolve('./src/index.ts'),
+            },
+        },
+    ];
 };
